--- frontend/server/src/app/(students)/lesson/[course_id]/[lesson_id]/[page]/style.css
@@ -1,48 +1,0 @@
-/* 教科書コンテンツ */
-.textbook {
-	padding: 10px;
-	line-height: 2;
-}
-
-/* 演習問題へのリンク */
-.box {
-	background: #c4d9ff;
-	border-left: #4e7bcc 5px solid;
-	padding: 10px;
-}
-.box p {
-	margin: 0;
-	padding: 0;
-}
-.indent_1 {
-	padding-left: 20px;
-}
-.indent_2 {
-	padding-left: 40px;
-}
-.solid_border_1 {
-	padding: 10px;
-	margin-bottom: 10px;
-	border: 1px solid #333333;
-}
-.dashed_border_1 {
-	padding: 10px;
-	margin-bottom: 10px;
-	border: 1px dashed #333333;
-}
-.example {
-	padding: 10px;
-	margin-bottom: 10px;
-	border: 2px dashed #333333;
-}
-.answer {
-	text-align: center;
-	width: 40px;
-	margin: 1em;
-	border: solid 2px #000000;
-}
-.definition {
-	padding: 10px;
-	margin-bottom: 10px;
-	border: 2px solid #333333;
-}
