+++ frontend/server/src/components/shared/MathJax.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import { MathJax as BetterMathJax, MathJaxContext } from "better-react-mathjax";
+import { MathJaxContext } from "better-react-mathjax";
 import type { FC, ReactNode } from "react";
 import ReactMarkdown from "react-markdown";
-// import rehypeMathjax from "rehype-mathjax";
+import rehypeMathjax from "rehype-mathjax";
 import rehypeRaw from "rehype-raw";
 import rehypeSanitize from "rehype-sanitize";
 import remarkGfm from "remark-gfm";
-// import remarkMath from "remark-math";
+import remarkMath from "remark-math";
 
 
 type Props = {
@@ -58,12 +58,9 @@ type MathJaxProps = {
 export const MathJax: FC<MathJaxProps> = (props) => {
   const { text } = props;
   return (
-    // <ReactMarkdown remarkPlugins={[remarkMath, remarkGfm]} rehypePlugins={[rehypeRaw, rehypeSanitize, rehypeMathjax]}>
-    <BetterMathJax>
-      <ReactMarkdown remarkPlugins={[remarkGfm]} rehypePlugins={[rehypeRaw, rehypeSanitize]}>
-        {text}
-      </ReactMarkdown>
-    </BetterMathJax>
+    <ReactMarkdown remarkPlugins={[remarkMath, remarkGfm]} rehypePlugins={[rehypeRaw, rehypeSanitize, rehypeMathjax]}>
+      {text}
+    </ReactMarkdown>
   );
 };
